requirements.txt

streamlit
google-generativeai
qiskit
qiskit-aer
pandas
